requirements.txt
absl-py==2.1.0
altair==4.1.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.2
dnspython==2.6.1
entrypoints==0.4
et-xmlfile==1.1.0
eventlet==0.36.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.4
Flask==3.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
flatbuffers==24.12.23
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.9.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.0
idna==3.7
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
mistune==3.0.2
ml_dtypes==0.5.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.4
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.6
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.0.0
pybase64==1.3.2
pycaw==20240210
pycparser==2.22
pydeck==0.9.0b1
pyforest==1.1.0
Pygments==2.17.2
Pympler==1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.2
pytz==2024.1
#pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.0
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
semver==3.0.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==70.1.0
simple-websocket==1.0.0
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.33.0
sympy==1.13.3
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.4.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.44.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
validators==0.34.0
virtualenv==20.26.0
watchdog==4.0.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.1
whitenoise==6.9.0
widgetsnbextension==4.0.11
wsproto==1.2.0
zipp==3.21.0